added migrations

+++ database/migrations/2019_11_01_01_create_tiers_table.php
@@ -15,7 +15,7 @@ class CreateTiersTable extends Migration
     {
         Schema::create('tiers', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->float('price', 6, 2);
+            $table->integer('price');
             $table->string('name');
             $table->integer('duration');
             $table->timestamps();
